src/flights/requirements.txt
aniso8601==9.0.1
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
db-dtypes==1.1.1
Flask==2.2.3
Flask-RESTful==0.3.9
google-api-core==2.11.0
google-auth==2.17.1
google-cloud-bigquery==3.9.0
google-cloud-core==2.3.2
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.59.0
grpcio==1.53.0
grpcio-status==1.53.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.24.2
packaging==23.0
pandas==2.0.0
proto-plus==1.22.2
protobuf==4.22.1
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.2
pytz==2023.3
requests==2.28.2
rsa==4.9
six==1.16.0
tzdata==2023.3
urllib3==1.26.15
Werkzeug==2.2.3
